First implementation of the train ETA on the PIDS

+++ src/main/java/mtr/data/RailwayData.java
@@ -109,7 +109,7 @@ public class RailwayData extends PersistentState {
 	}
 
 	public void simulateTrains(WorldAccess world) {
-		routes.forEach(route -> route.getPositionYaw(world, (int) world.getLunarTime()));
+		routes.forEach(route -> route.getPositionYaw(world, world.getLunarTime()));
 	}
 
 	// writing data
